Adição de classe para acesso as configurações de formatação.
